pyproject.toml
[project]
name = "customer-support-rag"
version = "0.1.0"
description = "End-to-end RAG system for customer support"
requires-python = ">=3.11"

dependencies = [
    "sentence-transformers>=2.7.0",
    "faiss-cpu>=1.8.0",
    "openai>=1.30.0",
    "datasets>=2.20.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "prometheus-client>=0.20.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "streamlit>=1.56.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "data",
    "indexing",
    "retrieval",
    "generation",
    "llmops",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
]
